refactor: move getByTextContent function into new testHelpers file
